:feat: 添加 stablediffusion api 适配
